Tile clearance only when hour is 3am

+++ index.php
@@ -109,9 +109,13 @@ if (!is_writable ($directory)) {
 $file = $cache . $layer . $location;
 file_put_contents ($file, $binary);
 
-# Clean out old files periodically
-if (rand (1, $garbageCollection1In) == 1) {
+# If the hour is 3am a clean out of old tiles may be triggered
+if (date('G') == 3 && rand (1, $garbageCollection1In) == 1) {
+	// A test of this on 22 Apr 2013 03:12:11 took five minutes to complete
+	error_log ('Starting tile clearance');
 	$command = "find {$_SERVER['DOCUMENT_ROOT']} -name '.png' -mtime +{$expiryDays} -exec rm -f {} \;";
+	$lastLine = exec ($command);
+	error_log ("Completed tile clearance: {$lastLine}");
 }
 
-?>
+# End of file
